AI SALES AUTOMATION/agency/lead-automation-system/backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
pandas==2.0.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.93.0
python-dotenv==1.0.0
requests==2.31.0
stripe==5.5.0
schedule==1.2.0
razorpay==1.4.1
